added pages and fixed layout
