+++ frontend/app/layouts/default.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="min-h-screen bg-gray-50 text-gray-800">
         <!-- organization name header -->
-        <header class="bg-blue-600 text-white text-center py-10 text-4xl font-bold">Admin</header>
+        <header class="bg-blue-600 text-white text-center py-10 text-4xl font-bold">Loading...</header>
         <Nav />
         <slot />
     </div>
