+++ frontend/app/pages/admin/courses.vue
