+++ frontend/app/pages/admin/index.vue
@@ -1,4 +1,7 @@
 <script lang="ts">
+definePageMeta({
+  layout: 'student-header'
+})
 export default defineNuxtComponent({
 
 })
