+++ frontend/app/pages/help.vue
@@ -1,4 +1,4 @@
-<script setup>
+<script lang="ts">
 definePageMeta({
   layout: 'student-header'
 })
