--- frontend/app/pages/student-records.vue
@@ -1,12 +1,0 @@
-<script setup>
-definePageMeta({
-  layout: 'student-header'
-})
-</script>
-
-<template>
-  <main class="p-6 text-center text-2xl">
-    place holder for student-records page
-  </main>
-</template>
-
